# feat(go/adbc/driver/snowflake): improve bulk ingestion speed