… docs/source/driver/snowflake.rst
… go/adbc/driver/snowflake/connection.go
… go/adbc/driver/snowflake/driver_test.go
… go/adbc/driver/snowflake/record_reader.go
… go/adbc/driver/snowflake/statement.go
… go/adbc/go.mod
… go/adbc/go.sum