@joellubi 2024-01-26
@zeroshade How do you feel about the resolution to changes you requested, and about handling the row group bytes written as a separate issue? I rebased to bring in the v15 libraries and was able to un-skip the test that depended on the issues fixed upstream. All appears to be green now.