@joellubi 2024-01-27
> @joellubi Looks good to me and we can address the bytes written as a separate issue. Thanks a ton for this! Would you mind commenting with any updated benchmarks on this comparing the old version to this, just so we have them in the notes (so we can include them in the release notes when this gets pushed out)

Thanks! I just re-ran the benchmarks on the current main branch in the same cloud environment as the existing benchmarks. The numbers came out basically the same so I would say the numbers in the PR description and in the documentation are accurate.